appending belonging spoken text
Search engine gives results in split spoken text. So when one person is constantly talking, the search enqine queries fragmented pieces of the whole. This results in inefficient scraping, because initially every result was individually loaded. Now, when pieces belong togehther, the spoken text is appended to each other and only once is a page loaded to seek the necessary information.